requirements.txt
Flask==3.0.3
gunicorn==22.0.0
requests==2.32.2
google-cloud-firestore==2.16.0
google-cloud-aiplatform==1.49.0
duckduckgo-search==6.3.2
google-generativeai>=0.7.2
flask-cors==4.0.1
